n2n 7.4 implementations

+++ src/app/n2n/core/config/ErrorConfig.php
@@ -36,7 +36,7 @@ class ErrorConfig {
 
 	public function __construct(bool $strictAttitude, bool $startupDetectErrors, bool $startupDetectBadRequests, 
 			bool $logSaveDetailInfo, bool $logSendMail, string $logMailRecipient = null, bool $logHandleStatusExceptions, 
-			array $logExcludedHttpStatuses, array $errorViewNames, private readonly ?int $monitorSlowQueryTime) {
+			array $logExcludedHttpStatuses, array $errorViewNames, private readonly ?float $monitorSlowQueryTime) {
 		$this->strictAttitude = $strictAttitude;
 		$this->startupDetectErrors = $startupDetectErrors;
 		$this->startupDetectBadRequests = $startupDetectBadRequests;
@@ -127,7 +127,7 @@ class ErrorConfig {
 		return $this->errorViewNames[self::ERROR_VIEW_DEFAULT_KEY_SUFFIX] ?? null;
 	}
 
-	function getMonitorSlowQueryTime(): ?int {
+	function getMonitorSlowQueryTime(): ?float {
 		return $this->monitorSlowQueryTime;
 	}
 }
